+++ platform/src/components/aws/index.ts
@@ -12,6 +12,7 @@ export * from "./cluster.js";
 export * from "./cognito-identity-pool.js";
 export * from "./cognito-user-pool.js";
 export * from "./cron.js";
+export * from "./cron-v2.js";
 export * from "./dns.js";
 export * from "./dynamo.js";
 export * from "./efs.js";
